fix(profile_sampler): cast denominator to double for mean_interval
Profile + clang build adds -Wsign-conversion -Wconversion under -Werror.
`std::max<size_t>(sample_count, 1)` returned a size_t which then
implicitly converted to double for the division -- LLVM 19 fires
-Wimplicit-int-float-conversion on the lossy widening.

Verified locally in docker (ubuntu-24.04 + clang-19.1.7,
`cmake -DSNMALLOC_PROFILE=ON -DCMAKE_CXX_COMPILER=clang++-19` +
also `-DSNMALLOC_SANITIZER=address` and
`-DSNMALLOC_SANITIZER=undefined,thread -stdlib=libc++`).

+++ src/test/func/profile_sampler/profile_sampler.cc
@@ -84,7 +84,7 @@ namespace
     const double total_bytes = static_cast<double>(N) * R;
     const double expected_samples = total_bytes / static_cast<double>(T);
     const double mean_interval =
-      total_bytes / std::max<size_t>(sample_count, 1);
+      total_bytes / static_cast<double>(std::max<size_t>(sample_count, 1));
 
     std::cout << "    N=" << N << " R=" << R << " T=" << T << "\n";
     std::cout << "    samples=" << sample_count << "  expected~"
